requirements.txt
flask==3.0.3
alibabacloud_iot20180120==5.1.0
python-dotenv==0.21.0

